fix: deduplicate the card definitions
to share across dashboards
